Move unit test string extension
